requirements_live.txt
# Live Trading Requirements
# =========================
# Install with: pip install -r requirements_live.txt

# Core
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Technical Analysis
ta>=0.10.0

# Interactive Brokers
ib_insync>=0.9.86

# Utilities
python-dateutil>=2.8.0
